Now building with nested parameters object.
