+++ java-to-zod-gradle-plugin/build.gradle.kts
@@ -16,6 +16,9 @@ dependencies {
     // https://mvnrepository.com/artifact/jakarta.validation/jakarta.validation-api
     implementation("jakarta.validation:jakarta.validation-api:3.1.0")
     implementation("sh.ivan:java-to-zod-core:0.7.0-SNAPSHOT")
+    // https://mvnrepository.com/artifact/org.apache.maven.plugin-tools/maven-plugin-annotations
+    compileOnly("org.apache.maven.plugin-tools:maven-plugin-annotations:3.15.0")
+
 
 
     // https://mvnrepository.com/artifact/org.projectlombok/lombok
@@ -33,6 +36,7 @@ dependencies {
     testImplementation("org.junit.jupiter:junit-jupiter-engine:5.8.1")
     testImplementation("org.junit.jupiter:junit-jupiter")
     testImplementation(gradleTestKit())
+
 }
 
 tasks.test {
